home_analytics/app/requirements.txt
duckdb==1.3.2
websocket-client==1.8.0

